App/backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database - MongoDB
motor==3.3.2
pymongo==4.6.0

# ML/Data Science
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.1.3
numpy==1.26.2

# File Handling
python-multipart==0.0.6
aiofiles==23.2.1

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0

# API Integrations
google-generativeai==0.8.6

# Additional Utilities
requests==2.31.0
httpx==0.25.2
